refactor(eval-author): move the Eval Author agent into Experimentalist
The 2026-07-28 extraction targeted a standalone Eval Author plugin. That goal
is retired: the customer-facing path is a skill rather than a plugin install,
and Experimentalist insight mode is the agent's only caller.

The agent sits next to the evaluator, staging, trace, and reporting helpers it
borrowed, which removes the package cycle that uv had to resolve and turns ten
rows of tracked debt into ordinary intra-package imports. The ratchet in
test_plugin_boundary.py drops from ten borrowed modules to one: the platform
client that discovery/ builds on.

The eval-author plugin keeps the `nemo agents eval-author` command group and
its discovery/ package. The move adds no dependencies, because Experimentalist
already declares harbor, nooa, pydantic, tomlkit, nemo-insights-plugin, and
nemo-platform-plugin. It drops nooa and tomlkit from the eval-author plugin,
which only the agent used.
