+++ plugins/nemo-eval-author/README.md
@@ -3,60 +3,66 @@
 
 # NeMo Eval Author Plugin
 
-Library-only plugin that owns the Eval Author agent (`eval_author/`).
+Owns the `nemo agents eval-author` command group, registered under `nemo.cli.agents` and
+mounted by the agents plugin.
+
+The Eval Author agent lives in the Experimentalist plugin, at
+[`nemo_experimentalist_plugin.eval_author`](../nemo-experimentalist/src/nemo_experimentalist_plugin/eval_author/README.md).
+Experimentalist insight mode is its only caller, so the agent sits next to the evaluator,
+staging, and trace helpers it depends on.
 
 ## Direction of travel
 
-**Eval Author is meant to become standalone, with nothing imported from
-Experimentalist.** Prefer duplicating a helper over sharing one, even when sharing
-looks tidier.
-
-Right now the two packages depend on each other:
+The dependency is one arrow. Eval Author borrows the platform client factory from
+Experimentalist, and Experimentalist imports nothing from here. Prefer duplicating a helper
+over adding a second arrow.
 
 | Arrow | Status | Why |
 | --- | --- | --- |
-| Experimentalist → Eval Author | permanent | insight mode imports `EvalAuthor` and `EvalAuthorConfig` at module scope |
-| Eval Author → Experimentalist | temporary | still borrows evaluator/Harbor, staging, trace, tools, cache, backend |
+| Eval Author → Experimentalist | one import | `discovery/run.py` builds a platform client with `make_client` |
+| Experimentalist → Eval Author | none | the agent moved into the Experimentalist plugin |
 
-[`tests/test_plugin_boundary.py`](tests/test_plugin_boundary.py) pins the second list
-so it can only shrink, and names what each remaining import is still for. `uv`
-resolves the current cycle; install both packages with:
+[`tests/test_plugin_boundary.py`](tests/test_plugin_boundary.py) pins the remaining import so
+it can only shrink.
+
+Install the plugin with:
 
 ```bash
 uv sync --group experimentalist
 ```
 
-## Public API
+## Discover
 
-```python
-from nemo_eval_author_plugin.eval_author.agent import EvalAuthor, build_eval_author_agent
-from nemo_eval_author_plugin.eval_author.models import EvalAuthorConfig, EvalAuthorResult
-from nemo_eval_author_plugin.eval_author.run import run_eval_author
+`nemo agents eval-author discover` validates one repository-owned Harbor config and writes no repository files.
 
-# Still borrowed from Experimentalist, and on the way out. Treat these as Eval Author's
-# own types once they move; do not build new code on the Experimentalist paths.
-from nemo_experimentalist_plugin.entities import Dataset, DatasetRef
-from nemo_experimentalist_plugin.experimentalist.components.dataset_staging import stage_eval_author_inputs
-from nemo_experimentalist_plugin.experimentalist.components.trace_analyzer import TraceAnalyzer
-from nemo_experimentalist_plugin.experimentalist.components.trace_explorer import TraceExplorer
-```
+The command accepts these flags:
 
-## Agent models
+- `--repo` selects the repository. The current directory is the default.
+- `--agent` sets the name. Without it, a string `agent` in root `optimizer.yaml` takes precedence over the repository directory slug.
+- `--dry-run` prints the report and uploads no files.
 
-Run `nemo setup` and select the default and fast models for the active Platform
-context. Eval Author uses the default model for authoring and the fast model for
-summarization. Press Enter at the fast-model prompt to reuse the default model.
+Discovery does not infer or generate a config. The preflight includes:
 
-The selections are workspace-qualified Platform Model Entity IDs. The Platform
-routes each request to the provider registered for that entity and reads its
-credential from Platform Secrets; Eval Author does not accept separate provider,
-endpoint, key, or model environment variables.
+- Harbor validates the schema, agent concurrency limits, task directories, and dataset coverage. Harbor also resolves the job.
+- Discovery identifies required environment variables. Harbor imports the agent and validates the environment backend.
+- `harbor job start --print-config` loads the config file.
 
-For non-interactive and isolated environments, `NEMO_DEFAULT_MODEL` and
-`NEMO_FAST_MODEL` can override the stored selections. Values must still use
-`workspace/model-name` and refer to Model Entities on the target Platform.
+The `ETHOS.md` and trace checks are advisory.
+Each invocation repeats the preflight and records one `sha256:<digest>` fingerprint plus the input file count. The fingerprint never skips validation.
 
-A `nemo agents eval-author` CLI is registered under `nemo.cli.agents` and
-mounted by the agents plugin. `discover` is implemented; `audit`, `propose`,
-`run`, and `doctor` remain placeholders. The library runner already uses the
-configured Platform model pair.
+Without `--dry-run`, the command uploads only `<agent>/discovery.md` to the `nemo-eval-author` fileset.
+Only a runnable report includes `cd <repo-root> && harbor job start -c <repo-relative-path>`.
+
+A standard run returns exit code 0 only for a runnable report and a successful upload.
+A dry run returns exit code 0 only if the report is runnable.
+All other outcomes return exit code 1. The command still uploads a report for an absent or rejected config.
+
+> **WARNING:** Use `discover` only with a trusted repository.
+> An agent import executes module top-level code.
+
+## Planned commands
+
+- `nemo agents eval-author audit` will report coverage gaps against `ETHOS.md` and will not change the current Harbor suite.
+- `nemo agents eval-author propose` will draft Harbor tasks and verifier patches for review.
+- `nemo agents eval-author run` will run `discover`, `audit`, and `propose` as one pipeline.
+- `nemo agents eval-author doctor` will check credentials, platform access, and the runtime for the other commands.
