+++ plugins/nemo-eval-author/pyproject.toml
@@ -4,19 +4,17 @@
 [project]
 name = "nemo-eval-author-plugin"
 version = "0.1.0"
-description = "Eval Author agent for NeMo Platform (hard-depends on Experimentalist for evaluator/trace helpers)."
+description = "Eval Author commands for NeMo Platform (borrows the Experimentalist platform client)."
 requires-python = ">=3.12,<3.14"
 dependencies = [
     "pydantic>=2",
     # Harbor 0.18 provides the discovery APIs used by this plugin.
     "harbor>=0.18",
-    "nooa>=0.0.9",
     "nemo-experimentalist-plugin",
     "nemo-insights-plugin",
     "nemo-platform",
     "nemo-platform-plugin",
     "pyyaml>=6.0.3",
-    "tomlkit>=0.13.3",
 ]
 
 [project.entry-points."nemo.cli.agents"]
