+++ plugins/nemo-experimentalist/src/nemo_experimentalist_plugin/eval_author/run.py
@@ -6,9 +6,9 @@
 from pathlib import Path
 from typing import Protocol
 
-from nemo_eval_author_plugin.eval_author.models import EvalAuthorConfig, EvalAuthorResult
 from nemo_experimentalist_plugin.client import make_client
 from nemo_experimentalist_plugin.entities import Dataset, DatasetRef, Task
+from nemo_experimentalist_plugin.eval_author.models import EvalAuthorConfig, EvalAuthorResult
 from nemo_experimentalist_plugin.experimentalist.components.dataset_staging import stage_eval_author_inputs
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.base import EvaluatorType
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.factory import DatasetFactory
@@ -148,7 +148,7 @@ def build_eval_author_agent(
     reporter: RunReporter | None = None,
 ) -> _EvalAuthorAgent:
     """Build the LLM-backed Eval Author agent lazily."""
-    from nemo_eval_author_plugin.eval_author.agent import build_eval_author_agent as _build_eval_author_agent
+    from nemo_experimentalist_plugin.eval_author.agent import build_eval_author_agent as _build_eval_author_agent
 
     return _build_eval_author_agent(
         experiment_dir=experiment_dir,
