+++ plugins/nemo-experimentalist/tests/eval_author/test_eval_author_materialization.py
@@ -10,9 +10,9 @@ import tomllib
 from pathlib import Path
 
 import pytest
-from nemo_eval_author_plugin.eval_author import materialization as materialization_module
-from nemo_eval_author_plugin.eval_author.materialization import InsightSuite, validate_metric_contracts
 from nemo_experimentalist_plugin.entities import Dataset, DatasetValidationError, Task
+from nemo_experimentalist_plugin.eval_author import materialization as materialization_module
+from nemo_experimentalist_plugin.eval_author.materialization import InsightSuite, validate_metric_contracts
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor import HarborDataset
