+++ plugins/nemo-experimentalist/tests/experimentalist/test_dataset_staging_runner.py
@@ -94,7 +94,7 @@ async def test_insight_run_stages_inputs_and_stops_at_eval_author_handoff(
         lambda: SimpleNamespace(build_evaluator=lambda *args, **kwargs: object()),
     )
     monkeypatch.setattr(runner_module, "DatasetFactory", RecordingDatasetFactory)
-    monkeypatch.setattr("nemo_eval_author_plugin.eval_author.agent.EvalAuthor", MutatingEvalAuthor)
+    monkeypatch.setattr("nemo_experimentalist_plugin.eval_author.agent.EvalAuthor", MutatingEvalAuthor)
 
     agent_dir = tmp_path / "agent"
     agent_dir.mkdir()
@@ -174,7 +174,7 @@ async def test_authored_metrics_and_suite_reach_the_run(
         runner_module, "EvaluatorFactory", lambda: SimpleNamespace(build_evaluator=lambda *a, **k: object())
     )
     monkeypatch.setattr(runner_module, "DatasetFactory", Factory)
-    monkeypatch.setattr("nemo_eval_author_plugin.eval_author.agent.EvalAuthor", Author)
+    monkeypatch.setattr("nemo_experimentalist_plugin.eval_author.agent.EvalAuthor", Author)
     monkeypatch.setattr(
         runner_module,
         "distribute_insight_suite_tasks",
