+++ pyproject.toml
@@ -71,9 +71,10 @@ version = "0.0.0"
 [dependency-groups]
 # Insights analyst plugin convenience group.
 insights = ["nemo-insights-plugin"]
-# Experimentalist and Eval Author currently support Python 3.12 and 3.13. Both are listed
-# because insight mode imports Eval Author at module scope, while Eval Author still depends
-# back on Experimentalist for evaluator/trace helpers (see plugins/nemo-eval-author/README.md).
+# Experimentalist and Eval Author support Python 3.12 and 3.13. Both are listed so the
+# `nemo agents eval-author` commands are installed alongside Experimentalist. The
+# dependency is one arrow: Eval Author borrows the Experimentalist platform client
+# (see plugins/nemo-eval-author/README.md).
 experimentalist = [
   "nemo-experimentalist-plugin ; python_full_version < '3.14'",
   "nemo-eval-author-plugin ; python_full_version < '3.14'",
